requirements.txt
requests==2.27.1
environs==9.5.0
pyTelegramBotAPI==4.12.*
python-telegram-bot==13.14
google-cloud-dialogflow==2.14.1
google-cloud-language==2.4.3
google-cloud-api-keys

